+++ tests/GUI-tests/GUI-tests.R
@@ -117,30 +117,33 @@ df <- data.frame(Question = c("", "", "Just a standard stem."), Option = options
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # expressions
-df <- data.frame(Question = c("", "", "Just a standard stem."), Option = options, Type = "radio",
+df <- data.frame(Question = c('', 'tags$h1("My header")', 'tags$b("This text is bold.")'), 
+                 Option = options, Type = "radio",
                  Stem = c('Math-stem.html', '', ''),
-                 StemExpression = c('', 'tags$h1("My header")', 'tags$b("This text is bold.")'))
+                 StemExpression = c(FALSE, TRUE, TRUE))
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
-df <- data.frame(Question = c("", "", ""), Option = options, Type = "radio",
-                 Stem = c('', '', ''),
-                 StemExpression = c('div(HTML("Here is <strong>one</strong> way to insert <em>arbitrary</em> HTML."))', 
-                                    'div(tags$style("#text { font-size: 35px; height: 200px; overflow: auto; }"), 
+df <- data.frame(Question = c('div(HTML("Here is <strong>one</strong> way to insert <em>arbitrary</em> HTML."))', 
+                              'div(tags$style("#text { font-size: 35px; height: 200px; overflow: auto; }"), 
                                          div(id = "text", paste(names(tags), collapse = ", ")))', 
-                                    'div(tags$style("#text { font-size: 20px; height: 65px; overflow: auto; }"), 
-                                         div(id = "text", paste(names(tags), collapse = ", ")))'))
+                              'div(tags$style("#text { font-size: 20px; height: 65px; overflow: auto; }"), 
+                                         div(id = "text", paste(names(tags), collapse = ", ")))'), 
+                 Option = options, Type = "radio",
+                 Stem = c('', '', ''),
+                 StemExpression = rep(TRUE, 3))
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # audio/video
 dirname <- paste0(getwd(), '/www')
 shiny::addResourcePath('www', dirname)
-df <- data.frame(Question = c("", "", "Just a standard stem."), Option = options, Type = "radio",
-                 Stem = c('Math-stem.html', '', ''),
-                 StemExpression = c('', 
-                                    'tags$audio(src = "www/clip.mp3", type = "audio/mp3",
+df <- data.frame(Question = c('', 
+                              'tags$audio(src = "www/clip.mp3", type = "audio/mp3",
                                     autoplay = TRUE, controls = TRUE)', 
-                                    'tags$video(src = "www/vid.mp4", type = "video/mp4",
-                                    controls = TRUE, height=260, width=260)'))
+                              'tags$video(src = "www/vid.mp4", type = "video/mp4",
+                                    controls = TRUE, height=260, width=260)'), 
+                 Option = options, Type = "radio",
+                 Stem = c('Math-stem.html', '', ''),
+                 StemExpression = c(FALSE, TRUE, TRUE))
 results <- mirtCAT(df = df, shinyGUI = list(forced_choice = FALSE))
 
 # checkbox input